+++ pkg/wms130/getfeatureinfo_kvp_request.go
