+++ pkg/wms130/getfeatureinfo_request.go
