+++ pkg/wms130/getmap_kvp_request.go
