+++ pkg/wmts100/capabilities_test.go
@@ -1,4 +1,4 @@
-package capabilities
+package wmts100
 
 import (
 	"encoding/xml"
